Move pools-related UI to their own package
