+++ features/storage/ui/src/main/kotlin/com/nasdroid/storage/ui/DI.kt
@@ -1,7 +1,7 @@
 package com.nasdroid.storage.ui
 
 import com.nasdroid.storage.logic.StorageLogicModule
-import com.nasdroid.storage.ui.overview.StorageOverviewViewModel
+import com.nasdroid.storage.ui.pools.overview.StorageOverviewViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.core.context.loadKoinModules
 import org.koin.dsl.module
