+++ features/storage/ui/src/main/kotlin/com/nasdroid/storage/ui/Navigation.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.nasdroid.storage.ui.overview.StorageOverviewScreen
+import com.nasdroid.storage.ui.pools.overview.StorageOverviewScreen
 
 /**
  * Registers a nested navigation graph for the Storage feature.
@@ -15,8 +15,8 @@ fun NavGraphBuilder.storageGraph(
     modifier: Modifier = Modifier,
     contentPadding: PaddingValues = PaddingValues()
 ) {
-    navigation(startDestination = "overview", route = route) {
-        composable("overview") {
+    navigation(startDestination = "poolsOverview", route = route) {
+        composable("poolsOverview") {
             StorageOverviewScreen(
                 modifier = modifier,
                 contentPadding = contentPadding,
